requirements.txt
docx2pdf==0.1.8
fpdf==1.7.2
langchain==0.3.25
langchain_community==0.3.24
langchain_neo4j==0.4.0
langchain_openai==0.3.19
marker==2.1.5
matplotlib==3.10.3
matplotlib_venn==1.1.2
neo4j==5.28.1
neo4j_graphrag==1.7.0
numpy==2.2.6
openai==1.84.0
openai_whisper==20240930
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
playsound==1.3.0
protobuf==6.31.1
PyAudio==0.2.14
pydub==0.25.1
PyQt6==6.9.0
PyQt6_sip==13.9.1
python-dotenv==1.1.0
pyttsx3==2.98
pywin32==308
Requests==2.32.3
SpeechRecognition==3.12.0
transformers==4.47.0
whisper==1.1.10
git+https://github.com/openai/whisper.git 